Build as monolithical shared library
